Add program comment favoriting
The program form is expanded to include comments if comments exists for the program's events. Users can favorite the comments they want displayed on the program overview and in the program report. The comments association is added to programs through events. The program report is refactored to use the new association. The favoriting display logic is refactored to use true booleans as opposed to a boolean that could potentially be null (facepalm).

This commit requires a rake db:migrate.